Number 7: Student watches porn and gets naked. A Bangledeshi 20 year old student was arrested after watching porn, stripping naked and attacking an attendant during a flight. In March 2018, shortly after the Malindo Air flight departed from Kuala Lumpur, the 20 year old started watching pornography on his laptop. As no one seemed to notice or call him out for it, the man began taking his clothes off. At the request of the cabin crew, the student put his clothes back on.
But soon, he tried to hug female flight attendants. When the crewmembers rejected him, he reportedly became aggressive, and groped a flight attendant. With the help of some of the passengers, the cabin crew managed to subdue the man, and tie his hands with a piece of cloth for the rest of the flight. One of the passengers managed to capture the moment with his phone. Although airline officials refuse to confirm details of the incident, the photo of a man with his bottom sticking out of his pants and hands tied up emerged online. The disruptive passenger was arrested on arrival.